code/requirements.txt

fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9


pydantic>=2.0


sqlalchemy>=2.0
bcrypt


pyjwt

litellm>=1.35.0
sentence-transformers
scikit-learn
vaderSentiment
numpy
pandas
langchain
langchain-community
langchain-huggingface


python-dotenv
httpx>=0.27.0
networkx
typer>=0.12.0
rich